feat(TASK-741): Add logging in BookController and add validation to id in deleteById method in BookService

+++ spring-data-jpa/src/main/java/ru/sabitov/example/service/BookService.java
@@ -50,6 +50,10 @@ public class BookService {
 
     @Transactional
     public void deleteById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new NotFoundException("Книга с id = %d не найдена".formatted(id));
+        }
+
         bookRepository.deleteById(id);
     }
